requirements.txt
python-telegram-bot>=21.0
pandas
numpy
sentence-transformers
scikit-learn
pandas
groq
torch